% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFS.R
\name{createIFS}
\alias{createIFS}
\title{CreateIFS is creating object which represents fractal to draw}
\usage{
createIFS(func_list)
}
\arguments{
\item{func_list}{list of functions to use}
}
\value{
fractal object
}
\description{
Function CreateIFS takes list of functions and creates fractal object
which can be later drawed with `plot` function
}
\examples{
func_list <- list(f1,f2)
createIFS(func_list)
}
