% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{getPoints}
\alias{getPoints}
\title{GetPoints for fractal}
\usage{
getPoints(.Object, point, depth)
}
\arguments{
\item{.Object}{IFS object}

\item{point}{starting point for drawing fractal}

\item{depth}{depth to which we want to call functions}
}
\description{
It makes data frame from points which we want to draw.
}
\examples{
func_list <- list(f1,f2)
fractal <- createIFS(func_list)
getPoints(fractal, c(0,0), 7)

}
