% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kochSnowflake.R
\docType{data}
\name{kochSnowflake}
\alias{kochSnowflake}
\title{kochSnowflake fractal}
\description{
Specification of koch snow flake fractal which can be plotted using plot function.
\href{https://en.wikipedia.org/wiki/Koch_snowflake}{Wikipedia}
}
\keyword{datasets}
