% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{plot}
\alias{plot}
\title{Plots given IFS object}
\usage{
plot(.Object, depth, color)
}
\arguments{
\item{.Object}{IFS object}

\item{depth}{depth to which we want to call functions}

\item{color}{optional color for fractal, default color is black}
}
\description{
Plots IFS object with given depth and color
}
\examples{
func_list <- list(f1,f2)
fractal <- createIFS(func_list)
plot(fractal, 6, 'red')

}
