% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythagoreanTree.R
\docType{data}
\name{pythagoreanTree}
\alias{pythagoreanTree}
\title{pythagoreanTree fractal}
\description{
Specification of pythagorean tree fractal which can be plotted using plot function.
\href{https://en.wikipedia.org/wiki/Pythagoras_tree_(fractal)}{Wikipedia}
}
\keyword{datasets}
