% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierpinskiCarpet.R
\docType{data}
\name{sierpinskiCarpet}
\alias{sierpinskiCarpet}
\title{Sierpinski carpet fractal}
\description{
Specification of sierpinski carpet fractal which can be plotted using plot function.
\href{https://en.wikipedia.org/wiki/Sierpinski_carpet}{Wikipedia}
}
\keyword{datasets}
