% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierpinskiTriangle.R
\docType{data}
\name{sierpinskiTriangle}
\alias{sierpinskiTriangle}
\title{Sierpinski triangle fractal}
\description{
Specification of Sierpinski triangle fractal with can be plotted using plot function.
\href{https://en.wikipedia.org/wiki/Sierpinski_triangle}{Wikipedia}
}
\keyword{datasets}
