% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{tree fractal}
\description{
Specification of tree fractal which can be plotted using plot function.
\href{https://en.wikipedia.org/wiki/Koch_snowflake}{Wikipedia}
}
\keyword{datasets}
