% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierpinski.R
\docType{data}
\name{sierpinskiTriangle}
\alias{sierpinskiTriangle}
\title{sierpinskiTriangle fractal}
\description{
Specification of sierpinskiTriangle fractal with can be plotted using plot function.
}
\keyword{datasets}
